feat(wpf): multi-class race window (#4 of 8, stage 2)
Extracts the console into a reusable RaceConsoleView UserControl and adds
the multi-class tab host that coordinates all classes — the WPF port of
MultiClassRaceForm.

- RaceConsoleView: the console content (header/body/action bar) as a
  UserControl with an IsHostedMode flag. Hosted mode suppresses the
  buyback + completion popups and routes Close to the parent; standalone
  mode keeps them. Persist saves the session and (hosted) the parent
  multi-class event. Explicit Teardown() avoids tab-switch unload races.
- RaceConsoleWindow: now a thin window hosting one RaceConsoleView
  (standalone / single-class).
- MultiClassRaceWindow: dark TabControl with one console per class.
  Coordinates the RR gate + buyback popup, tab-switch gating ("finish the
  round before switching"), per-class completion stats + popup, and the
  combined event summary. Tab state shown as a coloured dot
  (idle / started / next-up / RR-complete / done). Replays saved bracket
  state on load so interrupted events resume.
- TextSummaryWindow: dark scrollable summary dialog (avoids pulling
  WinForms into the WPF project).
- Landing now launches MultiClassRaceWindow for Start new event / Load,
  replacing the temporary first-class-only console launch.
