+++ src/RCDragManagerProd.WPF/Dialogs/TextSummaryWindow.xaml
@@ -1,0 +1,51 @@
+<Window x:Class="RCDragManagerProd.WPF.Dialogs.TextSummaryWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        Title="Summary"
+        Width="520" Height="560"
+        WindowStartupLocation="CenterOwner"
+        WindowStyle="None"
+        ResizeMode="CanResizeWithGrip"
+        KeyDown="Window_KeyDown"
+        Background="{StaticResource Brush.Background}"
+        FontFamily="Segoe UI"
+        ShowInTaskbar="False">
+
+    <WindowChrome.WindowChrome>
+        <WindowChrome CaptionHeight="38" GlassFrameThickness="0"
+                      ResizeBorderThickness="5" UseAeroCaptionButtons="False"/>
+    </WindowChrome.WindowChrome>
+
+    <Border Background="{StaticResource Brush.Background}">
+        <Grid>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="38"/>
+                <RowDefinition Height="*"/>
+                <RowDefinition Height="Auto"/>
+            </Grid.RowDefinitions>
+
+            <Grid Grid.Row="0" Background="{StaticResource Brush.Chrome}">
+                <TextBlock x:Name="LblTitle" Text="Summary"
+                           HorizontalAlignment="Center" VerticalAlignment="Center"
+                           FontSize="12" Foreground="{StaticResource Brush.TextHint}" FontWeight="Medium"/>
+                <StackPanel Orientation="Horizontal" HorizontalAlignment="Right"
+                            VerticalAlignment="Center" Margin="0,0,14,0">
+                    <Button Style="{StaticResource Style.WinCtrl.Close}" Click="BtnClose_Click"/>
+                </StackPanel>
+            </Grid>
+
+            <ScrollViewer Grid.Row="1" VerticalScrollBarVisibility="Auto" Padding="24,18">
+                <TextBlock x:Name="TxtBody"
+                           FontFamily="Consolas" FontSize="{StaticResource FontSize.Md}"
+                           Foreground="{StaticResource Brush.TextPrimary}"
+                           TextWrapping="Wrap"/>
+            </ScrollViewer>
+
+            <Border Grid.Row="2" Padding="24,12"
+                    BorderBrush="{StaticResource Brush.BorderSubtle}" BorderThickness="0,1,0,0">
+                <Button Style="{StaticResource Style.Button.Dialog.Primary}"
+                        Content="Close" Click="BtnClose_Click" HorizontalAlignment="Right"/>
+            </Border>
+        </Grid>
+    </Border>
+</Window>
