+++ src/RCDragManagerProd.WPF/Windows/MultiClassRaceWindow.xaml
@@ -1,0 +1,89 @@
+<Window x:Class="RCDragManagerProd.WPF.Windows.MultiClassRaceWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        Title="Multi-Class Event — RC Drag Manager"
+        Width="1180" Height="800"
+        MinWidth="980" MinHeight="640"
+        WindowStartupLocation="CenterScreen"
+        WindowStyle="None"
+        ResizeMode="CanResizeWithGrip"
+        Background="{StaticResource Brush.Background}"
+        FontFamily="Segoe UI">
+
+    <WindowChrome.WindowChrome>
+        <WindowChrome CaptionHeight="38" CornerRadius="0"
+                      GlassFrameThickness="0" ResizeBorderThickness="5"
+                      UseAeroCaptionButtons="False"/>
+    </WindowChrome.WindowChrome>
+
+    <Window.Resources>
+        <!-- Dark tab item -->
+        <Style x:Key="DarkTabItem" TargetType="TabItem">
+            <Setter Property="Foreground" Value="{StaticResource Brush.TextSecondary}"/>
+            <Setter Property="FontSize" Value="{StaticResource FontSize.Md}"/>
+            <Setter Property="Template">
+                <Setter.Value>
+                    <ControlTemplate TargetType="TabItem">
+                        <Border x:Name="bd" Background="{StaticResource Brush.Chrome}"
+                                Padding="16,9" Margin="0,0,2,0"
+                                BorderBrush="Transparent" BorderThickness="0,0,0,2"
+                                CornerRadius="6,6,0,0">
+                            <ContentPresenter ContentSource="Header"
+                                              HorizontalAlignment="Center" VerticalAlignment="Center"
+                                              TextElement.Foreground="{TemplateBinding Foreground}"/>
+                        </Border>
+                        <ControlTemplate.Triggers>
+                            <Trigger Property="IsSelected" Value="True">
+                                <Setter TargetName="bd" Property="Background" Value="{StaticResource Brush.Surface}"/>
+                                <Setter TargetName="bd" Property="BorderBrush" Value="{StaticResource Brush.Primary}"/>
+                                <Setter Property="Foreground" Value="{StaticResource Brush.TextPrimary}"/>
+                            </Trigger>
+                            <Trigger Property="IsMouseOver" Value="True">
+                                <Setter Property="Foreground" Value="{StaticResource Brush.TextPrimary}"/>
+                            </Trigger>
+                        </ControlTemplate.Triggers>
+                    </ControlTemplate>
+                </Setter.Value>
+            </Setter>
+        </Style>
+    </Window.Resources>
+
+    <Border Background="{StaticResource Brush.Background}">
+        <Grid>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="38"/>
+                <RowDefinition Height="Auto"/>
+                <RowDefinition Height="*"/>
+            </Grid.RowDefinitions>
+
+            <!-- Title bar -->
+            <Grid Grid.Row="0" Background="{StaticResource Brush.Chrome}">
+                <TextBlock x:Name="LblTitle" Text="Multi-class event"
+                           HorizontalAlignment="Center" VerticalAlignment="Center"
+                           FontSize="12" Foreground="{StaticResource Brush.TextHint}"
+                           FontWeight="Medium"/>
+                <StackPanel Orientation="Horizontal" HorizontalAlignment="Right"
+                            VerticalAlignment="Center" Margin="0,0,14,0">
+                    <Button Style="{StaticResource Style.WinCtrl.Minimize}" Click="BtnMinimize_Click" Margin="0,0,7,0"/>
+                    <Button Style="{StaticResource Style.WinCtrl.Maximize}" Click="BtnMaximize_Click" Margin="0,0,7,0"/>
+                    <Button Style="{StaticResource Style.WinCtrl.Close}" Click="BtnClose_Click"/>
+                </StackPanel>
+            </Grid>
+
+            <!-- Status strip -->
+            <Border Grid.Row="1" Background="{StaticResource Brush.Background}"
+                    Padding="20,6" MinHeight="14">
+                <TextBlock x:Name="LblStatus" Text=""
+                           FontSize="{StaticResource FontSize.Xs}"
+                           Foreground="{StaticResource Brush.Accent}"/>
+            </Border>
+
+            <!-- Tabs -->
+            <TabControl x:Name="Tabs" Grid.Row="2"
+                        Background="{StaticResource Brush.Background}"
+                        BorderThickness="0"
+                        ItemContainerStyle="{StaticResource DarkTabItem}"
+                        SelectionChanged="Tabs_SelectionChanged"/>
+        </Grid>
+    </Border>
+</Window>
